Added support for MassModuleStore event
This is a broader update than should have been needed.  MassModuleStore event breaks convention with the rest of the API, as it lacks the Localized version of the module names.  Realizing this, it makes much more sense to treat the internal names as the keys, and the Localized names just for output (duh).  I still needed to add in a display repair map, to fill in missing Localized types.